// src/lib/backend/api/api.ts
export interface V1AdminOrdersGet200Response {
    /**
     * 
     * @type {string}
     * @memberof V1AdminOrdersGet200Response
     */
    'code': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminOrdersGet200Response
     */
    'message': string | null;
    /**
     * 
     * @type {V1AdminOrdersGet200ResponseData}
     * @memberof V1AdminOrdersGet200Response
     */
    'data': V1AdminOrdersGet200ResponseData;
}
/**
 * 
 * @export
 * @interface V1AdminOrdersGet200ResponseData
 */
export interface V1AdminOrdersGet200ResponseData {
    /**
     * 
     * @type {Array<V1AdminOrdersGet200ResponseDataDataInner>}
     * @memberof V1AdminOrdersGet200ResponseData
     */
    'data': Array<V1AdminOrdersGet200ResponseDataDataInner>;
    /**
     * 
     * @type {number}
     * @memberof V1AdminOrdersGet200ResponseData
     */
    'count': number;
}
/**
 * 
 * @export
 * @interface V1AdminOrdersGet200ResponseDataDataInner
 */
export interface V1AdminOrdersGet200ResponseDataDataInner {
    /**
     * 
     * @type {string}
     * @memberof V1AdminOrdersGet200ResponseDataDataInner
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminOrdersGet200ResponseDataDataInner
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminOrdersGet200ResponseDataDataInner
     */
    'updatedAt': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminOrdersGet200ResponseDataDataInner
     */
    'deletedAt'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminOrdersGet200ResponseDataDataInner
     */
    'customerId': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminOrdersGet200ResponseDataDataInner
     */
    'currency'?: V1AdminOrdersGet200ResponseDataDataInnerCurrencyEnum;
    /**
     * 
     * @type {number}
     * @memberof V1AdminOrdersGet200ResponseDataDataInner
     */
    'totalPrice'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof V1AdminOrdersGet200ResponseDataDataInner
     */
    'totalDiscount'?: number | null;
}